Added Add and List functionality
