+++ src/main/java/Main.java
@@ -1,0 +1,15 @@
+public class Main {
+
+    public static void main(String[] args) {
+        Duke duke = new Duke();
+
+        String logo = " ____        _        \n"
+                + "|  _ \\ _   _| | _____ \n"
+                + "| | | | | | | |/ / _ \\\n"
+                + "| |_| | |_| |   <  __/\n"
+                + "|____/ \\__,_|_|\\_\\___|\n";
+        System.out.println("Hello from\n" + logo);
+
+        duke.run();
+    }
+}
